+++ RPI4/src/SensorHub.cpp
@@ -16,11 +16,11 @@ void SensorHub::ResetEnc(SensorHub::EncName encName)
 {
     SensorHub::GetInstance().ResetEncImpl(encName);
 }
-const std::array<u_int16_t, SensorHub::NUMENC> &SensorHub::GetEncData()
+const std::array<double, SensorHub::NUMENC> &SensorHub::GetEncData()
 {
     return std::ref(SensorHub::GetInstance().EncData);
 }
-const std::array<u_int16_t, SensorHub::NUMPRE> &SensorHub::GetPreData()
+const std::array<double, SensorHub::NUMPRE> &SensorHub::GetPreData()
 {
     return std::ref(SensorHub::GetInstance().PreData);
     // return std::cref(SensorHub::GetInstance().adc0.ReadData());
@@ -32,10 +32,10 @@ SensorHub::~SensorHub()
     munlockall();
 }
 SensorHub::SensorHub() //initialize member in list since Encoder has no default constructor
-    : LEncRecorder(Recorder<uint16_t,NUMENC/2>("EncodersL","LHipS,LKneS,LAnkS"))
-    , REncRecorder(Recorder<uint16_t,NUMENC/2>("EncodersR","RHipS,RKneS,RAnkS"))
-    ,PreRecorder(Recorder<uint16_t,NUMPRE>("Pressure","LTankPre,KnePre,Force,Pos,TankPre,na,na,na"))
-    ,PreRecOri(Recorder<u_int16_t,NUMPRE>("Pressure_ori","LTankPre,KnePre,Force,Pos,TankPre,na,na,na") )
+    : LEncRecorder(Recorder<double,NUMENC/2>("EncodersL","LHipS,LKneS,LAnkS"))
+    , REncRecorder(Recorder<double,NUMENC/2>("EncodersR","RHipS,RKneS,RAnkS"))
+    ,PreRecorder(Recorder<double,NUMPRE>("Pressure","LTankPre,KnePre,Force,Pos,TankPre,na,na,na"))
+    ,PreRecOri(Recorder<double,NUMPRE>("Pressure_ori","LTankPre,KnePre,Force,Pos,TankPre,na,na,na") )
     , LHipS_Enc(Encoder_L(Encoder_L::HIP1)), LKneS_Enc(Encoder_L(Encoder_L::KNEE)), LAnkS_Enc(Encoder_L(Encoder_L::ANK1)), RHipS_Enc(Encoder_R(Encoder_R::HIP1)), RKneS_Enc(Encoder_R(Encoder_R::KNEE)), RAnkS_Enc(Encoder_R(Encoder_R::ANK1)),adc0(ADC(0))//,adc1(ADC(1)) //, LHipF_Enc(Encoder_L(1)), LAnkF_Enc(Encoder_L(4)), RHipF_Enc(Encoder_R(1)), RAnkF_Enc(Encoder_R(4))
     ,filter_3_hz(FilterParam::Filter3Hz::a,FilterParam::Filter3Hz::b)
     ,filter_10_hz(FilterParam::Filter10Hz::a,FilterParam::Filter10Hz::b)
@@ -86,7 +86,7 @@ void SensorHub::UpdateLEnc()
     // senHub.EncData[SensorHub::LHipS] = senHub.LHipS_Enc.ReadPos();   //TODO: add robustness to ReadPos() when encoder is offline
     // senHub.EncData[SensorHub::LKneS] = senHub.LKneS_Enc.ReadPos();   //TODO: read the correct encoder when encoders connected
     // senHub.EncData[SensorHub::LAnkS] = senHub.LAnkS_Enc.ReadPos();
-    std::array<uint16_t,NUMENC/2> curMea{senHub.EncData[SensorHub::LHipS],senHub.EncData[SensorHub::LKneS],senHub.EncData[SensorHub::LAnkS]};
+    std::array<double,NUMENC/2> curMea{senHub.EncData[SensorHub::LHipS],senHub.EncData[SensorHub::LKneS],senHub.EncData[SensorHub::LAnkS]};
     
     senHub.LEncRecorder.PushData(curMea);
     // senHub.EncData[SensorHub::LHipF] = senHub.LAnkS_Enc.ReadPos();
@@ -100,17 +100,17 @@ void SensorHub::UpdateREnc()
     // senHub.EncData[SensorHub::RKneS]=senHub.RKneS_Enc.ReadPos();
     // senHub.EncData[SensorHub::RAnkS]=senHub.RKneS_Enc.ReadPos();
     // senHub.EncData[SensorHub::RAnkF]=senHub.RKneS_Enc.ReadPos();
-    std::array<uint16_t,NUMENC/2> curMea{senHub.EncData[SensorHub::RHipS],senHub.EncData[SensorHub::RKneS],senHub.EncData[SensorHub::RAnkS]};
+    std::array<double,NUMENC/2> curMea{senHub.EncData[SensorHub::RHipS],senHub.EncData[SensorHub::RKneS],senHub.EncData[SensorHub::RAnkS]};
     senHub.REncRecorder.PushData(curMea);
 }
 void SensorHub::UpdatePre()
 {
     SensorHub & senHub = SensorHub::GetInstance();
-    const std::array<u_int16_t,8> &data = senHub.adc0.ReadData();
+    const std::array<double,8> &data = senHub.adc0.ReadData();
 
     
 
-    std::array<u_int16_t,NUMPRE> cur_mea;
+    std::array<double,NUMPRE> cur_mea;
     cur_mea[0]=data[ADC::SEN0];
     cur_mea[1]=data[ADC::SEN1];
     cur_mea[2]=data[ADC::SEN2];
